refactor: move registry to runtime
