+++ src/main.rs
@@ -2,13 +2,11 @@ use arboard::Clipboard;
 use nannou::prelude::*;
 use nannou_egui::egui::{self, FontDefinitions, FontFamily};
 use nannou_egui::Egui;
-use once_cell::sync::Lazy;
 use std::cell::{Cell, RefCell};
-use std::collections::HashMap;
 use std::error::Error;
 use std::path::PathBuf;
 use std::sync::mpsc::Receiver;
-use std::sync::{mpsc, Mutex, Once};
+use std::sync::{mpsc, Once};
 use std::{env, fs, str};
 
 use framework::frame_controller;
@@ -22,96 +20,6 @@ mod sketches;
 
 const STORE_CONTROLS_CACHE_IN_PROJECT: bool = true;
 const GUI_WIDTH: u32 = 560;
-
-macro_rules! register_legacy_sketches {
-    ($registry:expr, $($module:ident),*) => {
-        $(
-            $registry.register(
-                &crate::sketches::$module::SKETCH_CONFIG,
-                |app, rect| {
-                    let model = crate::sketches::$module::init_model(
-                        app,
-                        WindowRect::new(rect)
-                    );
-                    Box::new(SketchAdapter::new(
-                        model,
-                        crate::sketches::$module::update,
-                        crate::sketches::$module::view,
-                    ))
-                }
-            );
-        )*
-    };
-}
-
-macro_rules! register_sketches {
-    ($registry:expr, $($module:ident),*) => {
-        $(
-            $registry.register(
-                &crate::sketches::$module::SKETCH_CONFIG,
-                |app, rect| {
-                    Box::new(crate::sketches::$module::init(
-                        app,
-                        WindowRect::new(rect)
-                    ))
-                }
-            );
-        )*
-    };
-}
-
-struct SketchInfo {
-    config: &'static SketchConfig,
-    factory: Box<
-        dyn for<'a> Fn(&'a App, Rect) -> Box<dyn Sketch + 'static>
-            + Send
-            + Sync,
-    >,
-}
-
-static REGISTRY: Lazy<Mutex<SketchRegistry>> =
-    Lazy::new(|| Mutex::new(SketchRegistry::new()));
-
-struct SketchRegistry {
-    sketches: HashMap<String, SketchInfo>,
-    sorted_names: Option<Vec<String>>,
-}
-
-impl SketchRegistry {
-    fn new() -> Self {
-        Self {
-            sketches: HashMap::new(),
-            sorted_names: None,
-        }
-    }
-
-    fn register<F>(&mut self, config: &'static SketchConfig, factory: F)
-    where
-        F: Fn(&App, Rect) -> Box<dyn Sketch> + Send + Sync + 'static,
-    {
-        self.sketches.insert(
-            config.name.to_string(),
-            SketchInfo {
-                config,
-                factory: Box::new(factory),
-            },
-        );
-    }
-
-    fn get(&self, name: &str) -> Option<&SketchInfo> {
-        self.sketches.get(name)
-    }
-
-    fn names(&mut self) -> &Vec<String> {
-        if self.sorted_names.is_none() {
-            let mut names: Vec<String> =
-                self.sketches.keys().cloned().collect();
-            names.sort();
-            self.sorted_names = Some(names);
-        }
-        self.sorted_names.as_ref().unwrap()
-    }
-}
 
 fn main() {
     init_logger();
