+++ src/runtime/mod.rs
@@ -1,4 +1,5 @@
 pub mod gui;
 pub mod prelude;
 pub mod recording;
+pub mod registry;
 pub mod shared;
