+++ src/runtime/prelude.rs
@@ -1,3 +1,4 @@
 pub use crate::runtime::gui::*;
 pub use crate::runtime::recording::*;
+pub use crate::runtime::registry::*;
 pub use crate::runtime::shared::*;
